lagd slettModal for destinasjoner

+++ WebApp_Mappe2/ClientApp/src/app/destinasjon/destinasjonListe/destinasjoner.component.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { destinasjon } from "../../destinasjon";
 import { Router } from '@angular/router';
+import { Modal } from '../../slettModal';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'app-root',
@@ -11,8 +13,9 @@ import { Router } from '@angular/router';
 export class DestinasjonComponent {
   public alleDestinasjoner: Array<destinasjon>;
   public laster: string;
+  destTilSletting: string;
 
-  constructor(private http: HttpClient, private router: Router) { }
+  constructor(private http: HttpClient, private router: Router, private modalService: NgbModal) { }
 
   ngOnInit() {
     this.hentAlleDestinasjoner();
@@ -25,21 +28,41 @@ export class DestinasjonComponent {
         this.alleDestinasjoner = data;
         this.laster = "";
       },
-      error => alert(error),
-      () => console.log("Alle destinasjoner har blitt hentet.")
-    );
+        error => alert(error),
+        () => console.log("Alle destinasjoner har blitt hentet.")
+      );
   }
 
-  slettDestinasjon(id: number){
-    this.http.delete('api/Destinasjon/' + id)
-      .subscribe(retur => {
-        
+  slettDestinasjon(id: number) {
+    this.http.get<destinasjon>('api/Destinasjon/' + id)
+      .subscribe(destinasjon => {
+        this.destTilSletting = destinasjon.sted + " - " + destinasjon.land;
+        this.modalSlett(id);
+
       },
-        error => console.log(error),
-        () => console.log("Sletting av id:  " + id + " gjennomført.")
-    );
-    this.hentAlleDestinasjoner();
-    this.router.navigate(['/destinasjonListe']);
+        error => console.log(error)
+      );
+  }
+
+  modalSlett(id: number) {
+
+    const modalRef = this.modalService.open(Modal);
+
+    modalRef.componentInstance.navn = this.destTilSletting;
+
+    modalRef.result.then(retur => {
+      console.log('Lukket med: ' + retur);
+      if (retur == "Slett") {
+
+        this.http.delete('api/Destinasjon/' + id)
+          .subscribe(retur => {
+            this.hentAlleDestinasjoner();
+          },
+            error => console.log(error),
+            () => console.log("Sletting av id:  " + id + " gjennomført.")
+          );
+      }
+      this.router.navigate(['/destinasjonListe']);
+    });
   }
 }
-
